backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.3.0
sqlmodel>=0.0.22
ollama>=0.2.0

# Document parsing
pypdf>=4.2.0
python-docx>=1.1.2

